+++ src/main/java/insane96mcp/iguanatweaksreborn/module/client/Misc.java
@@ -2,7 +2,6 @@ package insane96mcp.iguanatweaksreborn.module.client;
 
 import insane96mcp.iguanatweaksreborn.module.ClientModules;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
@@ -15,22 +14,19 @@ import net.minecraftforge.client.gui.overlay.VanillaGuiOverlay;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Misc", description = "Misc client side changes")
-@LoadFeature(module = ClientModules.Ids.CLIENT)
+@LoadFeature(module = ClientModules.Ids.CLIENT,
+        name = "Misc",
+        description = "Misc client side changes")
 public class Misc extends Feature {
-    @Config
-    @Label(name = "No tilting with non-directional damage types", description = "If true, camera will not tilt when taking magic, wither, on fire, cramming, drowning and thorns damage.")
+    @Config(description = "If true, camera will not tilt when taking magic, wither, on fire, cramming, drowning and thorns damage.")
     public static Boolean noTiltingWithNonDirectionalDamageTypes = true;
 
-    @Config
-    @Label(name = "Red block outline with wrong tool", description = "If true, the outline around blocks will be red if the tool in hand will make drops not ... drop.")
+    @Config(description = "If true, the outline around blocks will be red if the tool in hand will make drops not ... drop.")
     public static Boolean redBlockOutlineWithWrongTool = true;
 
-    @Config(min = 0)
-    @Label(name = "Floaty hotbar", description = "Moves the hotbar this amount of pixels up (like bedrock edition). Other mods' GUI elements should work flawlessly if using the correct Forge GUI fields")
+    @Config(min = 0, description = "Moves the hotbar this amount of pixels up (like bedrock edition). Other mods' GUI elements should work flawlessly if using the correct Forge GUI fields")
     public static Integer floatyHotbar = 2;
-    @Config
-    @Label(name = "Fix mounts GUI", description = "If true, hunger will be rendered event when the player is in riding an entity. Also hides the jump bar unless jumping (showing the XP bar)")
+    @Config(description = "If true, hunger will be rendered even when the player is in riding an entity. Also hides the jump bar unless jumping (showing the XP bar)")
     public static Boolean fixMountsGui = true;
 
     public Misc(Module module, boolean enabledByDefault, boolean canBeDisabled) {